notifications started

+++ src/Components/Student/StudentsLogs/EditPracticeLog.js
@@ -1,6 +1,7 @@
 import React from 'react'
 import { connect } from 'react-redux'
 import {editLog, getLog, deleteLog} from '../../../redux/reducers/logsReducer'
+import {getAllLogsForStudent} from '../../../redux/reducers/studentReducer'
 
  class EditPracticeLog extends React.Component{
     constructor(){
@@ -31,8 +32,10 @@ import {editLog, getLog, deleteLog} from '../../../redux/reducers/logsReducer'
         }
     }
 
-    saveChanges = () => {
+    saveChanges = (e) => {
         const {log_date, log_time, log_material} = this.state
+
+        e.preventDefault()
 
         let editedLog = {
             log_date,
@@ -40,32 +43,47 @@ import {editLog, getLog, deleteLog} from '../../../redux/reducers/logsReducer'
             log_material
         }
 
-        this.props.editLog(this.props.match.params.id, editedLog)
-        this.props.getLog(this.props.match.params.id)
+        if(window.confirm('Are you sure you want to edit this practice log?') === true){
+    
+            this.props.editLog(this.props.match.params.id, editedLog)
+            this.props.getLog(this.props.match.params.id)
+    
+            this.setState({
+                log_date:'',
+                log_time:'',
+                log_material:''
+            })
+    
+            alert('Practice log has been updated!')
 
-        this.setState({
-            log_date:'',
-            log_time:'',
-            log_material:''
-        })
+            this.props.getAllLogsForStudent(this.props.student.student_id)
 
-        alert('Practice log has been updated!')
-
-        this.props.history.push('/student/logs')
+            this.props.history.push('/student/logs')
+        }
     }
 
     deleteLog = () => {
 
-        this.props.deleteLog(this.props.match.params.id)
+        if(window.confirm('Are you sure you want to delete this practice log?') === true){
+            this.props.deleteLog(this.props.match.params.id)
+    
+            this.setState({
+                log_date:'',
+                log_time:'',
+                log_material:''
+            })
 
-        this.setState({
-            log_date:'',
-            log_time:'',
-            log_material:''
-        })
+            this.props.getAllLogsForStudent(this.props.student.student_id)
 
-        this.props.history.push('/student/logs')
+            this.props.history.push('/student/logs')
+
         }
+
+    }
+
+    backToCreate = () => {
+        this.props.history.push('/student/logs')
+    }
 
 
     handelInputChange = (e) => {
@@ -79,13 +97,14 @@ import {editLog, getLog, deleteLog} from '../../../redux/reducers/logsReducer'
         return(
             <div>
                 <div>EditPracticeLog</div>
-                <div>
-                    <input name='log_date' value={this.state.log_date} onChange={this.handelInputChange} placeholder='Date'/>
-                    <input name='log_time' value={this.state.log_time} onChange={this.handelInputChange} placeholder='Minutes Practiced'/>
-                    <input name='log_material' value={this.state.log_material} onChange={this.handelInputChange} placeholder='What did you practice?'/>
-                    <button onClick={this.saveChanges}>Save Changes</button>
-                    <button onClick={this.deleteLog}>Delete Practice Log</button>
-                </div>
+                    <form onChange={this.handelInputChange} onSubmit={this.saveChanges}>
+                        <input name='log_date' value={this.state.log_date || ''} onChange={this.handelInputChange} type='datetime-local' require='true'/>
+                        <input type='number' name='log_time' min='5' step='5' value={this.state.log_time || ''} onChange={this.handelInputChange} placeholder='Minutes Practiced' require='true'/>
+                        <textarea name='log_material' value={this.state.log_material || ''} onChange={this.handelInputChange} placeholder='What did you practice?' require='true'/>
+                        <input type='submit' value='Submit Practice Log'/>
+                        <button type='button' onClick={this.deleteLog}> Delete Practice Log</button>
+                    </form>
+                    <button onClick={this.backToCreate}>Back</button>
             </div>
         )
     }
@@ -97,4 +116,4 @@ const mapStateToProps = (reduxState) => {
         logs: reduxState.logsReducer.logs[0]
     }
 }
-export default connect(mapStateToProps, {editLog, getLog, deleteLog})(EditPracticeLog)
+export default connect(mapStateToProps, {editLog, getLog, deleteLog, getAllLogsForStudent})(EditPracticeLog)
